Add manifest number to check status api response
